0437 fixing

+++ BackTestingFinal/MetricVar.cs
@@ -10,9 +10,9 @@ namespace BackTestingFinal
     {
         public bool isLong;
 
-        public double CR = 1D;
+        public decimal CR = 1m;
         public double Kelly = 1D;
-        public List<double> ProfitRates = new List<double>();
+        public List<decimal> ProfitRates = new List<decimal>();
 
         public DrawDownData DD = default;
         public DrawDownData MDD = default;
